Revert "first"
This reverts commit 9c868869c040015def27e613791dfb0886848c9c.

# Conflicts:
#	src/main/java/com/example/demo/model/User.java

+++ src/main/java/com/example/demo/model/User.java
@@ -8,17 +8,17 @@ import jakarta.persistence.*;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     private String name;
     private String email;
 
     // Getters and Setters
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
